+++ app/src/main/java/woowacourse/movie/list/view/ReservationHistoryFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.list.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,10 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import woowacourse.movie.databinding.FragmentSettingBinding
+import woowacourse.movie.R
+import woowacourse.movie.databinding.FragmentReservationHistoryBinding
 
-class SettingFragment : Fragment() {
-    private lateinit var binding: FragmentSettingBinding
+class ReservationHistoryFragment : Fragment() {
+    private lateinit var binding: FragmentReservationHistoryBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -17,7 +18,10 @@ class SettingFragment : Fragment() {
         savedInstanceState: Bundle?,
     ): View {
         binding =
-            DataBindingUtil.inflate(inflater, R.layout.fragment_setting, container, false)
+            DataBindingUtil.inflate(
+                inflater,
+                R.layout.fragment_reservation_history, container, false,
+            )
 
         return binding.root
     }
