+++ app/src/main/java/woowacourse/movie/list/view/SettingFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.list.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,10 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import woowacourse.movie.databinding.FragmentReservationHistoryBinding
+import woowacourse.movie.R
+import woowacourse.movie.databinding.FragmentSettingBinding
 
-class ReservationHistoryFragment : Fragment() {
-    private lateinit var binding: FragmentReservationHistoryBinding
+class SettingFragment : Fragment() {
+    private lateinit var binding: FragmentSettingBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -17,7 +18,7 @@ class ReservationHistoryFragment : Fragment() {
         savedInstanceState: Bundle?,
     ): View {
         binding =
-            DataBindingUtil.inflate(inflater, R.layout.fragment_reservation_history, container, false)
+            DataBindingUtil.inflate(inflater, R.layout.fragment_setting, container, false)
 
         return binding.root
     }
