+++ app/src/main/res/layout/activity_home.xml
@@ -7,7 +7,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".HomeActivity">
+        tools:context=".list.view.HomeActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragment"
