+++ app/src/main/res/layout/fragment_reservation_history.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".ReservationHistoryFragment">
+        tools:context=".list.view.ReservationHistoryFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
