+++ app/src/main/res/layout/fragment_setting.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".SettingFragment">
+        tools:context=".list.view.SettingFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
