+++ app/src/main/res/layout/fragment_theater.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".TheaterFragment">
+        tools:context=".list.view.TheaterFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:background="@color/white"
